Download must-gather diagnostics - server side
Currently, in the case of an issue, the customer needs to figure out
on their own where they need to go to get the diagnostics data of
each applicable component. It would be a major User Experience
improvement to have this automated and fully integrated into the UI.
This enhacement indicate how server side should be implemented.