[image-registry] Automate Image Pruning
Automate the image pruning maintenance task for the registry.
This is a necessary addition to ensure an OpenShift cluster remains
healthy.
Image pruning shall be configured by a new cluster-scoped CRD
which defines the pruner's configuration.
The image registry operator will then install the pruner and report
appropriate status, metrics, and alerts.