Add audit logs to cluster-logging-log-forwarding enhancement proposal
This is an addition to the already accepted proposal for log forwarding,
which suggests the addition of the audit logs in a separate pipeline.

These logs will not be forwarded by default.

This comes from the need to store audit logs in a secure location, as
part of custmers needing to meet government regulations.