enhancements: Standardize YAML vs. header title syntax
The previous docs were ambiguous, which lead to divergent
implementations and confusion about what was supposed to be lowercased
and dashed.  This commit makes the template more specific and tries to
standardize the existing enhancements.  But if there is a clear rule
about how to process the Markdown header, I'm not clear on why we need
a YAML title property at all.