enhancements/oc/mirroring-configuration-manifests: Propose a new enhancement
Describing how we're going to feed restricted-network clusters the
release-signature ConfigMaps that the cluster-version operator reads
since openshift/cluster-version-operator#279.

Personally, I prefer 'manifests' naming here, so:

* manifests directory
* --apply-manifests and --manifests-to-dir options.

But there was some concern that the Kubernetes/OpenShift manifests
would be confused with the container image manifests in
v2/openshift/release/manifests [1].  kubernetes-manifests was
suggested [2], which would to a kubernetes-manifests directory and
--apply-kubernetes-manifests and --kubernetes-manifests-to-dir
options.

But there was some concern about using Kubernetes anywhere [3], so the
commit goes with a config directory and --apply-config and
--config-to-dir options.

The truncated digest hash in the example signature filename avoids any
concerns about filename length limitations on restrictive operating
systems or filesystems [4].

[1]: https://github.com/openshift/enhancements/pull/238#discussion_r388271776
[2]: https://github.com/openshift/enhancements/pull/188#discussion_r393734352
[3]: https://github.com/openshift/enhancements/pull/188#discussion_r395756071
[4]: https://github.com/openshift/enhancements/pull/188#discussion_r393737361