ClusterLogForwarder enhancement proposal.
This is the (close to) final draft of the ClusterLogForwarder proposal for GA.
It replaces the tech-preview LogForwarding API.

Several requirements have been removed since the last draft to reduce the GA
scope.  They are still relevant for later releases, and are outlined in "Future
Plans" at the end of the proposal.

Signed-off-by: Alan Conway <aconway@redhat.com>