enhancements/update/operators: Collect operator-dev-doc
Move the previous operator-dev-doc down into this document as a
README, and drop a reference in the old location for backwards-compat
with existing linkers.  I'd prefer a symlink, but GitHub's tree
browser just renders the link path as an unclickable string [1].

In the spirit of 28c6ca7e3b (enhancements/update: Consolidate
update-related enhancements, 2020-08-05, #422).

[1]: https://github.com/dear-github/dear-github/issues/156#issuecomment-222349916