Template: add specific guidance in upgrade/downgrade section
Capture some more specific guidance on upgrade/downgrade expectations
to help authors think through their scrnarios.

In https://bugzilla.redhat.com/1794360 we see an example of debate
around our expectations for handling downgrades and rollbacks. The
conclusions there are good examples of the kind of information that is
useful to add to the template.