network: on premise Service load balancers
OpenShift does not currently support Services of type=LoadBalancer on
bare metal or other on premise infrastructure environments.  This
enhancement proposes a way forward, which is to adopt MetalLB.

The document explores some high level requirements, discusses some
alternatives considered, and maps out how we would get to work on
this through technical due diligence, upstream community engagement,
and careful planning of OpenShift integration with a new operator.

There is not currently a target for when this would be fully
supported, as this is a proposed enhancement on a direction to take.
More extensive technical due diligence, development, and testing will
help define the roadmap over time.