Improve oc awareness of ImageContentSourcePolicy
This proposal is to add logic in `oc` to look for ImageContentSourcePolicy from a cluster or from a file to be used in `oc adm release` commands.
If using a mirrored image and the mirrored source registry is disconnected, `oc adm release extract|mirror` do not succeed. This is because
the mirrored image tags (the individual component images from a payload) retain references to the mirrored registry, usually something like
`quay.io/openshift-release-dev/ocp-v4.0-art-dev`, and `oc adm release` only looks up the image reference that a disconnected user would not have access to.
Implementing this proposal will solve this issue. `oc` should look for ICSP in the cluster/current context if connected,
if the user has permission to access ICSPs, and if the current flow of using the image-reference from a given image fails.
When a cluster context is unclear or there's the possibility that the connected cluster could be a different target than expected,
`oc` should not default to connecting to the cluster.