template: Add reviewers and approvers guidance
While writing an enhancement, it occurred to me that there was not any
written guidance on who should be specified as reviewers of approvers.
This is my attempt to capture what seems to be done already to help
others understand who to ask for approval.  It leaves a lot of room
for interpretation, but avoids introducing any additional unnecessary
process overhead.