Enhancement for adding UnqualifiedSearchRegitries to the Image CR
Users want the ability to configure the list of registries that are
used when pull an image using short name. This field already exists
in the /etc/containers/registries.conf file and we want to expose this
via the images.config.openshift.io cluster CR so users can easily configure
this list in the same spot they do for allowed, blocked, and insecure registries.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>