Discuss host network configuration interfaces
This change introduces a new enhancement that discusses host network
configuration.  It is different from the typical enhancement in that
its goal is informational and to discuss what is already present in
this area.  It also provides references to other related works in
progress.

I wrote this first for myself and propose it here in hopes that it may
help others.  I find that this context is important and helpful to
understand when discussing or reviewing enhancements for a specific
feature related to host network configuration.

I imagine this as a living document that should get updated as key
improvements are made to how we manage host network configuration for
OpenShift.