metallb: Describe how load balancing works
Despite being called MetalLB, the project does not actually implement
load balancing as it is typically thought of (something like haproxy).
It is one piece of a larger setup that results in load balancing for
Service Load Balancers.  These additions to the doc attempt to help
explain further how load balancing works with MetalLB.