enhancements/update: Consolidate update-related enhancements
These are all related to the cluster-version operator, which is the
update team.  Collect them all together in a single directory to make
them easier to find.  I didn't bother leaving "there used to be an
enhancement here, but it has moved to..." markers in the old
locations, because I expect traffic is low, but we could do that if it
turns out folks had linked the old locations and can't find the new
ones.

Source for "51 excluded manifests in 4.3.0":

  $ oc adm release extract --to manifests quay.io/openshift-release-dev/ocp-release:4.3.0-x86_64
  Extracted release payload from digest sha256:3a516480dfd68e0f87f702b4d7bdd6f6a0acfdac5cd2e9767b838ceede34d70d created at 2020-01-22T02:05:15Z
  $ grep -hr exclude.release.openshift.io manifests | sort | uniq -c
       51     exclude.release.openshift.io/internal-openshift-hosted: "true"