network: Add a high level BGP use cases overview
Sometimes when "BGP" is mentioned, it's not always clear which use case
someone is talking about.  This is a complex topic.  BGP itself is
complex and so is the discussion of how it might apply to an OpenShift
environment.

This document aims to start capturing notes on different use cases we
envision to help build some common understanding and common langauge for
future discussions.