Add ingress fault detection proposal
This enhancement proposal outlines
an extension to the OpenShift Ingress Operator
that gives the Ingress Operator the ability to alert
OpenShift users when Cluster Ingress appears to be
malfunctioning. This enhancement proposal also outlines the
details of an Ingress "debug" container that can be invoked
on an OpenShift cluster to pinpoint potential Ingress issues.