template: Add an informational enhancement status
I've started writing some enhancements that are meant more to discuss
the context for a problem space and not to propose any specific changes.
One example is an overview of how BGP relates to OpenShift, and a list
of use cases that each may be explored in their own future enhancements.
Another is a doc that discusses the different interfaces related to
configuring networking on a host (see PR #394).

These sorts of docs don't fit the normal status workflow, so I propose
adding a new "informational" state that covers these cases.