Enhancement for User Space Pod Interface and API Library
This enhancement describes the changes needed to enable running user space
applications (i.e. DPDK based applications) running in a pod on an OCP
deployment.

The pod needs additional information to be deployed properly, like hugepage
information and interface mapping. This enhancement describes how the data
will be passed to pod, but focuses on an API Library that is intended to be
used by CNF developers to consume the provided data.

Signed-off-by: Billy McFall <22157057+Billy99@users.noreply.github.com>