add a tool to report on activity in this repository
Add a tool to generate an activity report for enhancements to be used
as a newsletter for people interested in following the work happening
here but who do not have the time to follow every review individually.

The default output includes details for pull requests closed or with
active discussion in the last week, as well as some other lists such as
very stale reviews that might need attention or to be closed.

Use markdown format, to make converting the output to HTML email and
static pages (such as #476) easier.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>