Add weekly newsletters to the repository
Doug Hellmann (@dhellmann) and I have started experimenting with a
weekly newsletter digest of activity in the OpenShift enhancements
repository.  This commit includes the editions from the first few weeks
we've tried it.

The bulk of the structure of the newsletter is generated by a report
tool that pulls data from the Github API.  We're still iterating on the
details of the data collected and its output, but the tool will come in
a future PR to this repository.

Adding these directly to github will make them easier to discover and
refer back to than only having them in a mailing list archive or
something similar.