Add single-node-developer Cluster Profile
Add a new ‘single-node-developer’ cluster profile that defines the set
of OpenShift payload components, and the configuration thereof,
applicable to single-node resource-constrained non-production OCP
clusters that support the development of OpenShift applications to be
deployed for production on other OCP clusters, such as those using the
default ‘self-managed-high-availability’ profile. The
‘single-node-developer’ profile will be defined for and utilized in
producing the Code Ready Containers (CRC) product that runs in a
single VM on a developer’s workstation or laptop. This profile is not
intended or supported for any kind of production deployment.