New method for providing configurable self-hosted LB/DNS/VIP stack for on-prem
In current implementation,  the self-hosted DNS/LB/VIP stack runs under the MCO umbrella by default
in all on-prem platforms and there's no option to configure it.
Some customers usually have their own external Load Balancing and DNS resolution.
Besides, there are cases in which some traffic used to provide these self-hosted services
is disallowed in the customer's network.

The outcome of such scenarios is that cluster's resources are spent on providing unused
services.
This PR suggests a new configurable method for providing the self-hosted stack.