update machine-api-usage-telemetry
During implementation of this enhancement we have determined that there
will be a need for two labels on the MachineHealthCheck related
metrics. Although these labels have unbounded cardinality, it will only
need to be exported through telemetry as a sum without the need for the
labels.

The alert for machines with old deletion timestamps has been renamed to
`MachineNotYetDeleted`, to better describe the root cause as opposed to
the symptom.