cluster capabilities api
As we add more deployment topologies for OpenShift, we need a way to
communicate expectations to the operators managing the cluster. This
enhancement describes an API for expressing the expected
"capabilities" of a cluster, including aspects such as how hard
operators should work to provide high-availability and whether
operators should configure their operands in a way to minimize
resource consumption for especially constrained environments.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>