OLM Managed Operators Max OpenShift Versions
This commit introduces a proposal that enables OLM managed operators to
specify a maximum OpenShift version that they support. OLM can then use
this information to prevent minor OpenShift upgrades when an installed
operator will not run on the next minor OpenShift version. Additionally,
OLM can use this information so it does not install an operator on an
unsupported OpenShift version.