convert from github/super-linter to markdownlint
Our github org does not support actions, so we need to be able to run
the lint job in prow. This change adds a Dockerfile to define an image
with the tool, changes the Makefile to support using that image on a
local developer system, and updates the config file to adjust some of
the additional issues found by running the tool differently.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>