move the rules for sorting prs into groups up to main
This is the first step to refactoring the PR report tool so that it is
easier to add new groups organized in more flexible ways. This moves
the logic for sorting items into the "buckets" up to the main program,
so the stats package can be used for different types of reports.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>