fix group detection for non-enhancement prs
When trying to determine the group from the list of files changed in
the PR we were ignoring files that were not markdown entries. This
made all tools, hack, and housekeeping changes use the "general" group
and show up in the output.

This change uses all filenames and adds more explicit rules for some
of the housekeeping files

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>