tools: fix summary formatting
Stop trying to unwrap paragraphs or otherwise clean up formatting,
since the linter job enforces that in our input files now. Simply
prefix each line with the formatting needed to turn it into a block
quote, leaving all of the other text unmodified so it renders using
the nested structure.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>