OKD: Move proposal to `implementable`, clarify summary
Moves the OKD proposal to `implementable` to reflect OKD's present state of being
continuously released in a rolling fashion.

In the summary, clarifies the goals of OKD and adds a paragraph that clarifies that
implementing and maintaining support for OKD in individual components shall not require
new levels of commitment by engineers and shall not be considered blocking when working
on OCP-related items.