tools: read accounts to ignore from config file
Extend the configuration file format to include a place to put the
list of github accounts to ignore when building reviewer stats. This
makes it easy for users to add common bot accounts to their
configuration file, instead of specifying them every time the tool
runs. If the --ignore command line option is used, any values from the
command line override the settings in the config file.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>