hack/template-lint: Actually, use PULL_BASE_SHA
Instead of 320ec318e3 (hack/template-lint: Diff from origin/master,
2021-04-26, #755), to avoid [1]:

  fatal: ambiguous argument 'origin/master..': unknown revision or path not in the working tree.
  Use '--' to separate paths from revisions, like this:
  'git <command> [<revision>...] -- [<file>...]'

PULL_BASE_SHA is part of Prow [2].

When folks call us outside of Prow, fall back to origin/master.  This
isn't a great fallback for folks using the script in local
development, because they may not have called the upstream remote
'origin', but:

  $ PULL_BASE_SHA=openshift/master hack/template-lint.sh

or whatever is possible, if a bit of a hack.  We considered HEAD, but
that's not a great match for folks looking to adjust an existing local
commit after noticing their pull request failing the presubmit.

Also add the '--' suggested in Git's error message, to show that we
intend this to be a revision ID and not a file path.

[1]: https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/pr-logs/pull/openshift_enhancements/744/pull-ci-openshift-enhancements-master-markdownlint/1386818564115140608/artifacts/test/build-log.txt
[2]: https://github.com/kubernetes/test-infra/blob/73e6e4703b38b43fe262ba0554d642e9aebcc0e8/prow/jobs.md#job-environment-variables

squash! hack/template-lint: Actually, use PULL_BASE_SHA

origin/master