Drop Node Hostname Resolution
In some Edge deployment scenarios, compute resources on remote
worker nodes can be in short supply. Deployers of these have
objected to the presence of unnecessary services on their nodes.
As a result, we would like to remove any services that are no
longer required for the cluster to function.