Proposed OLM-based Monitoring Stack Solution for RH Managed Services and future needs.
As an effect of the collaboration of Monitoring Team with Service Delivery org in our Monitoring Enhancement Working Group, with @fpetkovski would like to propose solution for immdiate needs for Managed Services (SaaS/Layered Services/Addons).

Feedback very welcome! Please keep all offline discussions in this PR so everyone is up-to-date.

Signed-by: Bartlomiej Plotka <bwplotka@gmail.com>