Create Network Observability enhancement document
Network Observability introduces a new category in OpenShift that
provides networking information for a single cluster. It gives insight
into what's on the network, when and what types of traffic and traffic
flows are being made, and by whom. It gathers data to help design, plan,
and answer questions about the network and provides a visual
representation to help understand, diagnose, and troubleshoot networking
issues.