tools: allow the upstream git remote to have any name
Instead of requiring that the local clone 'origin' point to the
openshift/enhancements fork, allow any name. Look for the remote with
the right URL when it is needed. Avoid looking for the remote for
every PR by creating a new Summarizer type that holds the remote
information in a way that can be reused for most operations that need
it, other than the initial repository configuration.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>