CSI Inline Ephemeral Volume Security
This enhancement proposes a new admission plugin to check that the CSI
driver used to provision a `csi` volume has been vetted and deemed
"safe" for a given PodSecurity profile. Components which install
CSIDriver drivers can optionally add a label to the driver object, which
declares that the driver is safe for a given PodSecurity profile. The
admission checks if a pod references a CSI volume on creation, and if
so, looks up the referenced CSIDriver object and checks the value of the
`security.openshift.io/csi-ephemeral-volume` label. If the driver's
profile is more permissive than the PodSecurity profile of the pod's
namespace, admission is denied.