Enhancement for improving image pull via shortname security
This is a proposal to improve the security of image pulls
when the image is refered to by shortname. This is already
implemented in podman and buildah and we would like to
extend this to cri-o so that OCP can benefit from it.

There is no API change happening here. Creating an enhancement
so that we can get feedback on the approach here and whether
this would be helpful for OpenShift. Would also like to get approval
on the list of images in the alias table to be shipped with OCP once
this enhancement is approved.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>