Dual Stack VIPs for on-prem deployments
Originally the on-prem loadbalancer architecture supported only single-stack
ipv4 or ipv6 deployments. When dual stack support was added for the baremetal
platform, we carried over the VIP configuration that only allowed for a single
VIP to be specified. The thinking was that in dual stack clusters every node
would have access to either ipv4 or ipv6 endpoints. While this is true of the
nodes, it is not necessarily true of the applications that run in the cluster
and may need to interact with the API or ingress services.

This design is proposing the addition of a second VIP for dual stack clusters
so API and ingress are accessible from either ip version.