Automated Workflow for Agent-based Installer
Summary

"Cluster 0" deployments, for the first cluster in an environment, are
unique because they occur in situations where there may not be a lot
of other infrastructure to support long-running services normally
associated with automated provisioning for on-premises use
cases. Nevertheless, users and partners want to automate these
deployments. The assisted installer GUI provides an excellent user
experience for manually deploying clusters. This enhancement covers
the automation use case.