enhancements/update/update-blocker-lifecycle: Pivot to "update recommendations"
Users are occasionally confused by the "blocked edges" wording, which
is internal shorthand.

The fact that update recommendations are tracked as a directed graph
(hence "edges") isn't all that relevant for most users, so talk about
"updates" instead of "edges".

While we occasionally drop update recommendations to reduce the
chances that users update into serious issues, from [1]:

  When an update recommendation is supported, it remains supported for
  the life of 4.10, even if the update recommendation is later dropped
  or made conditional.

Updating is still possible, we just don't think it's a good idea
(unless maybe you're elbow deep in a support case, and a support tech
decides to recommend an update based on context that is not available
to the update service).  So replace "blocked" with "not recommended",
to convey the "still possible" aspect.

Also ask for some more specifics in the impact statement template, as
we gear up to support conditional updates fff9d3c349
(enhancements/update/targeted-update-edge-blocking: Propose a new
enhancement, 2020-08-07, #821).

[1]: https://docs.openshift.com/container-platform/4.10/updating/understanding-upgrade-channels-release.html#upgrade-version-paths_understanding-upgrade-channels-releases