Automated Workflow for Agent-based Installer (#1067)
* Automated Workflow for Agent-based Installer

Summary

"Cluster 0" deployments, for the first cluster in an environment, are
unique because they occur in situations where there may not be a lot
of other infrastructure to support long-running services normally
associated with automated provisioning for on-premises use
cases. Nevertheless, users and partners want to automate these
deployments. The assisted installer GUI provides an excellent user
experience for manually deploying clusters. This enhancement covers
the automation use case.

* use 3 level command structure

change `agent wait-for-install-complete` to `agent wait-for install-complete`

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* highlight that this is case relies on user-provided automation

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* remove open question about feature flag

We will develop in a feature branch, but that detail isn't relevant to
the design in the way that a feature flag would be, so it's not
mentioned in the enhancement now that we've answered the question.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* remove qualifier in non-goal for collecting telemetry

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>

* don't say "minimum viable" version

* replace hyphen with space in installer sub-command names

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>