MGMT-10403: SNO + Workers enhancement - adjust upgrade non-goal to be a goal
The non-goal in question is:

```markdown
Deal with expansion of clusters that have been installed before the
implementation of this enhancement (if possible, their expansion may be
addressed by an upgrade followed with documentation outlining the steps
that the user has to take to enable expansion, e.g. patch the default
`IngressController` to have node selector targetting the master pool
rather than the worker pool).
```

Having it be a non-goal seems to have been a mistake, as it's really
easy to solve and would save a lot of headache in terms of complexity
for users upgrading from 4.10 to 4.11 - we wouldn't have to document
the differences between a "true freshly installed 4.11" cluster and
a "cluster upgraded to 4.11", expanding it with workers should work
just the same.

Implementation of this adjustment:

https://github.com/openshift/cluster-ingress-operator/pull/767