kubernetes for industrial edge
This enhancement describes MicroShift, which provides an essential
container orchestration runtime compatible with kubernetes and
OpenShift, but built for Internet-of-things (IoT) and Edge computing
scenarios that are both CPU and memory constrained.  The container
orchestration runtime is binary compatible with OpenShift Container
Platform, but it is not 100% API resource compatible.  It has a strict
subset of OpenShift API resources pertinent for IoT and Edge computing
scenarios with a strong bias to retaining traditional runtime security

Quite a lot of the text is taken from other working documents.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>