tools: install node 14 differently on fedora and rhel
The base image used for local developer testing is fedora, but the
image used in CI is a RHEL 8 image. Use the steps from the
nodesource.com script for enabling their RPM repo so we can install
from there in CI, and use the Fedora version when building the image
locally.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>