dev-guide/cluster-fleet-evaluation: Mention conditional update risks
And also conditional Insights gathering.  Because backporting a change
to existing 4.y and then waiting for the fleet to update can take a
while.  But both of these conditional approaches use Red Hat hosted
APIs to deliver query logic to clusters dynamically, so they can get
some cluster state out of existing clusters without having to backport
anything or wait for customers to update.

Conditional update risks also allow developers who are backporting
some in-cluster Prometheus metric, including, but not limited to
cluster_operator_conditions{condition="ClusterFleetEvaluation"}, to
dynamically give (or remove) update advice based on that metric,
without having to commit to particular advice before backporting.
Although, when we are confident enough to be able to commit to advice
when backporting, using something like an Upgradeable condition that
has defined in-cluster semantics allows those semantics to also apply
to restricted-network clusters that may not be using a local OpenShift
Update Service [1].

[1]: https://docs.openshift.com/container-platform/4.11/updating/updating-restricted-network-cluster.html#update-restricted-network-cluster-update-service