automatically detect the last date a report was generated
Instead of requiring the caller to set DAYSBACK properly, look for the
last report file present in the repo and use that to figure out how
long ago the report was generated.

Move all of the logic for doing that out of the Makefile into a script
where it is easier to read and debug.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>