update external provider enhancement
this change updates a name from the code sample to be consistent with
the created code. When reviewing the code for the API implementation of
the external platform type, we had a realization about the variable name
we are using to distinguish the provider supplied name for the
infrastructure. For more information, see the discussion on original
pull request[0].

[0] https://github.com/openshift/api/pull/1301