Support for Multiarch Manager Operator
This enhancement proposes a multiarch manager operator which will be a
secondary OLM managed operator installable through OperatorHub. The
operator will expose functionality which when configured, would allow
pods to be scheduled on the appropriate architecture nodes based on the
architecture of the container images.

The name of the operator is chosen as such in case future functionality
would need to be added that relates to multi-architectural workings.

xref: https://issues.redhat.com/browse/MIXEDARCH-215

Co-authored-by: Alessandro Di Stefano <aleskandro@redhat.com>