enhancements/update/targeted-update-edge-blocking: Drop Evaluating
Following up on 6af8d09143
(enhancements/update/targeted-update-edge-blocking: Update after API
landed, 2021-09-20, #910), with more changes to align this enhancement
with the actual implementation.

We never implemented Evaluating, because Recommended=Unknown means
we're trying (and failing) to figure out if some of the target's risks
apply to the cluster.

And:

  // +kubebuilder:validation:Required
  // +kubebuilder:validation:MinItems=1
  // +listType=atomic
  // +required
  MatchingRules []ClusterCondition `json:"matchingRules"`

means that empty matchingRules is not possible, as previously discussed around:

  Mitigating that "where did that edge go?" confusion in the
  required-property alternative, David expects graph-admins to be
  keeping watch over enough ClusterVersion content...

in the alternatives section of the enhancement.