Shared Resources via OpenShift Builds Operator
Proposal to remove the Shared Resource CSI driver from OpenShift,
and deploy it via the OpenShift Builds operator instead. The
primary purpose of this document is to solicit feedback on this
approach. The CSI driver is currently marked as Tech Preview, and as
such does not require a proposal for its removal.

The `CSIDriverSharedResource` feature in OCP will be marked as
deprecated in 4.15, with the intent of removal in 4.16. Clusters
which want to add this CSI driver will need to install the OpenShift
Builds operator first. Once installed, an admin will need to enable
the driver through the operator's configuration custom resource.

Alternatives considered include keeping the CSI driver in OCP and
marking it GA, or deploying the CSI driver through its own operator
and corresponding layered product. The proposal details the disadvantage
of these approaches, and why removal from OCP was chosen as the path
forward.

Signed-off-by: Adam Kaplan <adam.kaplan@redhat.com>