dev-guide/host-port-registry: Add process and control-plane only columns
"control plane only" had shown up in a number of notes, but it wasn't
being consistently answered.  I've added a new column to structure
that information, so it's clear whether an entry is control-plane
only, not control-plane only, or not yet declared.

I've also added a new protocol column, because folks have been asking
what protocol the cluster-version operator uses for serving metrics.
With the new column, it's clear that the CVO is using HTTPS.  Folks
who maintain other components will need to go through and fill in the
new columns for their entries, although I've guessed where the nodes
seemed clear.