enhancements/update/channel-rename-generally-available: New enhancement
Trying to package [1] as an enhancement to explore motivation, risks,
alternatives, and impacted components.  The Feature itself does not
currently spend much time on motivation, so most of that is me
guessing from my own experience, and not necessarily a reflection on
why the business folks want to perform this rename.

[1]: https://issues.redhat.com/browse/OCPSTRAT-1153