Drop CRD defaulting
Use runtime validation by the ocm-o cluster operator to handle
defaulting of the new `spec.builderServiceAccount` field. This allows
us flexibility to change the default behavior of OCP in the future.
There is a minor cost to implementing this in the operator, instead of
CRD defaulting mechanisms.

The proposal was also updated to use the new enhancement template.