Add support for performance profile status in hypershift
Up until now, there was no defined resolution over how PerformanceProfile
status conditions with the status of the different components created
from the PerformanceProfile (MachineConfig, KubeletConfig, Tuned) are handled and exposed.
In this PR, we suggest adding a custom ConfigMap that will be used as a middleware object.
It will hold the updated status populated by the performance profile controller
and will be watched by the NodePool controller, which will calculate a brief overview
of the current status provided and reflect it under NodePool.status.condition.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>